Trust Region Methods: move to separate namespace
