+++ src/Numerics/Optimization/TrustRegion/TrustRegionDogLegMinimizer.cs
@@ -1,4 +1,4 @@
-﻿namespace MathNet.Numerics.Optimization
+﻿namespace MathNet.Numerics.Optimization.TrustRegion
 {
     public sealed class TrustRegionDogLegMinimizer : TrustRegionMinimizerBase
     {
